refactor to add aerosandbox.numpy
