+++ aerosandbox/numpy/arithmetic.py
@@ -1,9 +1,9 @@
 import numpy as np
 import casadi as cas
-from aerosandbox.optimization.math.array import length
+from aerosandbox.numpy.array import length
 
 
-def sum1(x):
+def sum(x):
     """Returns the sum of a vector x."""
     try:
         return np.sum(x)
@@ -13,4 +13,4 @@ def sum1(x):
 
 def mean(x):
     """Returns the mean of a vector x."""
-    return sum1(x) / length(x)
+    return sum(x) / length(x)
