+++ aerosandbox/numpy/array.py
