+++ aerosandbox/numpy/calculus.py
