+++ aerosandbox/numpy/logicals.py
